Debugging Calibration
Ran the TEMPer windows exe through a disassembler/decompiler (JetBrains
dotPeek); It turns out it's a .Net application and can be pulled to bits
trivially (It's written in C#).
This shows us what the actual app is doing, so we *should* be able to
get the values we decode to match those from the app (which are
presumably the same as those you get from pressing the button).